Refactor: renovate frontend and backend architecture

+++ backendemailservice/src/main/java/com/backendemailservice/backendemailservice/dto/EmailActionRequestDto.java
@@ -3,14 +3,15 @@ package com.backendemailservice.backendemailservice.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+
 public class EmailActionRequestDto {
 
     @NotNull(message = "Email ID must not be null")
     @Positive(message = "Email ID must be a positive number")
-    private Integer emailId;
+    private Long emailId;
 
     public EmailActionRequestDto() {}
 
-    public Integer getEmailId() { return emailId; }
-    public void setEmailId(Integer emailId) { this.emailId = emailId; }
+    public Long getEmailId() { return emailId; }
+    public void setEmailId(Long emailId) { this.emailId = emailId; }
 }
