+++ backendemailservice/src/main/java/com/backendemailservice/backendemailservice/dto/EmailResponseDto.java
@@ -2,9 +2,10 @@ package com.backendemailservice.backendemailservice.dto;
 
 import java.time.LocalDateTime;
 
+
 public class EmailResponseDto {
 
-    private Integer emailID;
+    private Long emailID;
     private String sender;
     private String receiver;
     private String subject;
@@ -13,7 +14,7 @@ public class EmailResponseDto {
     private LocalDateTime date;
     private boolean trash;
 
-    public EmailResponseDto(Integer emailID, String sender, String receiver, String subject, String body, String priority, LocalDateTime date, boolean trash) {
+    public EmailResponseDto(Long emailID, String sender, String receiver, String subject, String body, String priority, LocalDateTime date, boolean trash) {
         this.emailID = emailID;
         this.sender = sender;
         this.receiver = receiver;
@@ -24,7 +25,7 @@ public class EmailResponseDto {
         this.trash = trash;
     }
 
-    public Integer getEmailID() { return emailID; }
+    public Long getEmailID() { return emailID; }
     public String getSender() { return sender; }
     public String getReceiver() { return receiver; }
     public String getSubject() { return subject; }
